refactor(inventory-transaction): align domain model and implement DTOs (Phase 21a)
- Apply standard parsePositiveInt validation in parseInventoryTransaction
- Refactor stock types into dedicated files: material-stock.ts and storage-location-stock.ts
- Update barrel index for inventory-transactions
- Standardize stock parsing logic with better safety guards
